+++ Assets/ViveGrip/Scripts/Internal/ViveGrip_Highlighter.cs
@@ -2,8 +2,7 @@ using UnityEngine;
 using System.Collections.Generic;
 
 public class ViveGrip_Highlighter : MonoBehaviour {
-  Color tintColor = Color.white;
-  float tintRatio = 0.2f;
+  private Color tintColor = new Color(0.2f, 0.2f, 0.2f);
   private Queue<Color> oldColors = new Queue<Color>();
 
   void Start () {}
@@ -13,7 +12,8 @@ public class ViveGrip_Highlighter : MonoBehaviour {
     foreach (Material material in GetComponent<Renderer>().materials) {
       Color currentColor = material.color;
       oldColors.Enqueue(currentColor);
-      material.color = Color.Lerp(currentColor, color, tintRatio);
+      Color tintedColor = currentColor + color;
+      material.color = tintedColor;
     }
   }